+++ app/src/main/java/ch/epfl/skysync/database/tables/BasketTable.kt
@@ -1,9 +1,14 @@
 package ch.epfl.skysync.database.tables
 
+import ch.epfl.skysync.database.DateLocalDateConverter
 import ch.epfl.skysync.database.FirestoreDatabase
 import ch.epfl.skysync.database.Table
 import ch.epfl.skysync.database.schemas.BasketSchema
+import ch.epfl.skysync.models.calendar.TimeSlot
 import ch.epfl.skysync.models.flight.Basket
+import ch.epfl.skysync.models.flight.Flight
+import com.google.firebase.firestore.Filter
+import java.time.LocalDate
 
 /** Represent the "basket" table */
 class BasketTable(db: FirestoreDatabase) :
@@ -17,6 +22,32 @@ class BasketTable(db: FirestoreDatabase) :
    */
   suspend fun add(item: Basket, onError: ((Exception) -> Unit)? = null): String {
     return withErrorCallback(onError) { db.addItem(path, BasketSchema.fromModel(item)) }
+  }
+
+  /**
+   * Returns the available baskets on a given date and timeslot
+   *
+   * @param localDate The requested day
+   * @param timeslot The requested timeslot
+   * @param onError Callback called when an error occurs
+   */
+  suspend fun getBasketsAvailableOn(
+      flightTable: FlightTable,
+      localDate: LocalDate,
+      timeslot: TimeSlot,
+      onError: ((Exception) -> Unit)?
+  ): List<Basket> {
+    val dateFilter = Filter.equalTo("date", DateLocalDateConverter.localDateToDate(localDate))
+    val timeslotFilter = Filter.equalTo("timeSlot", timeslot)
+    val flightFilter = Filter.and(dateFilter, timeslotFilter)
+
+    val unavailableBasketsIds: Set<String> =
+        flightTable
+            .query(flightFilter, onError)
+            .mapNotNull { flight: Flight -> flight.basket?.id }
+            .toSet()
+
+    return getAll(onError).filterNot { basket: Basket -> basket.id in unavailableBasketsIds }
   }
 
   companion object {
