+++ app/src/main/java/ch/epfl/skysync/database/tables/UserTable.kt
@@ -1,12 +1,16 @@
 package ch.epfl.skysync.database.tables
 
+import ch.epfl.skysync.database.DateLocalDateConverter
 import ch.epfl.skysync.database.FirestoreDatabase
 import ch.epfl.skysync.database.Table
 import ch.epfl.skysync.database.schemas.UserSchema
 import ch.epfl.skysync.models.calendar.Availability
+import ch.epfl.skysync.models.calendar.AvailabilityStatus
+import ch.epfl.skysync.models.calendar.TimeSlot
 import ch.epfl.skysync.models.flight.Flight
 import ch.epfl.skysync.models.user.User
 import com.google.firebase.firestore.Filter
+import java.time.LocalDate
 import kotlinx.coroutines.async
 import kotlinx.coroutines.awaitAll
 import kotlinx.coroutines.coroutineScope
@@ -55,6 +59,60 @@ class UserTable(db: FirestoreDatabase) : Table<User, UserSchema>(db, UserSchema:
               .awaitAll()
       flights.filterNotNull()
     }
+  }
+
+  /**
+   * Returns the available user on the given day and timeslot
+   *
+   * @param localDate The requested day
+   * @param timeslot The requested timeslot
+   * @param onError Callback called when an error occurs
+   */
+  suspend fun getUsersAvailableOn(
+      flightTable: FlightTable,
+      localDate: LocalDate,
+      timeslot: TimeSlot,
+      onError: ((Exception) -> Unit)?
+  ): List<User> = coroutineScope {
+    val dateFilter = Filter.equalTo("date", DateLocalDateConverter.localDateToDate(localDate))
+    val timeslotFilter = Filter.equalTo("timeSlot", timeslot)
+    val dateTimeSlotFilter = Filter.and(dateFilter, timeslotFilter)
+
+    // Retrieve all flights of the given day, timeslot
+    val flightsIds =
+        flightTable.query(dateTimeSlotFilter, onError).map { flight: Flight -> flight.id }
+
+    // Retrieve all possible available members
+    var potentialAvailableUsers = getAll(onError)
+
+    // If there are flights on the given day
+    // Retrieve all members of these flights (they are not available)
+    if (flightsIds.isNotEmpty()) {
+      val unavailableUserIds =
+          flightMemberTable.query(Filter.inArray("flightId", flightsIds), onError).map { fm ->
+            fm.userId
+          }
+      // Remove these user from the possible available members
+      // Now there are only user left who can be available on the given day and timeslot
+      potentialAvailableUsers =
+          potentialAvailableUsers.filterNot { user: User -> user.id in unavailableUserIds }
+    }
+    // Return all user who are available on the given day and timeslot
+    return@coroutineScope potentialAvailableUsers
+        .map { user: User ->
+          async {
+            val availableUsers =
+                availabilityTable.query(
+                    Filter.and(Filter.equalTo("userId", user.id), dateTimeSlotFilter))
+
+            if (availableUsers.firstOrNull()?.status == AvailabilityStatus.OK) {
+              return@async user
+            }
+            return@async null
+          }
+        }
+        .awaitAll()
+        .filterNotNull()
   }
 
   /** Remove user from all its flight assignments */
